1.0.7 Channel Names Update
A lot of people have been asking for a way to update channel names, and I have been saying no for awhile. The reason why I didn't want to do this in the first place is because many people will run into the rate limit set by discord, and claim the bot doesn't update frequent enough.

The rate limit set by discord is 2 requests per 10 minutes of time, and I cannot change that, but hey! It does the thing at least some of the time! :100: :+1: