CFX Update
Simplified project layout.
Update to .Net Core 6

Provider's are class that return required data to update the bots status (IServerInformationProvider)

To start, there are two providers available.

SteamProvider = 0
CFXProvider = 1

Depending on which provider is chosen then the difference in logic will execute, but provide the same resulting data. Goal is to expand to other types later.