Service Update
Created the concept of Services, and Providers in the application.

A service is an outside application that providers data to accomplish updating the status typically a REST api.
A provider is the logic required to fetch, and generate the required data into a response object.

There are 4 Provider Types:
0 - Steam
1 - CFX
2 - Scum
3 - Minecraft

Prevent spamming of exceptions in Provider logic will prevent the spam during server restarts.
Added HttpExecuter logic to project