Test Cases and Docker Fix
After the last update, I had left an issue with the parsing of docker configuration variables for Application variables. This fixes the possible null, the option of having steam or battle metrics, or none.

I have built test cases around this parsing logic, so that in the future if I break something I know before releasing it. Please note that this issue doesn't extend to the standard configuration as JSON hold a specific format for dictionaries that would cause the parser to fail.

I also moved the logic for setting channel name to an extension method.