Rotating Status Update
There was a request to make it possible to define many statuses, and have the bot rotate through them each update. This wasn't a huge change, but I wanted to at least break this off as the first release after 1.1.0 as there were already so many changes in that version.

This commit changes the usage of StatusFormat to use a helper methods to get the current format, or null.
Documentation still needs to be added.
Added docker-compose comment for status usage.