Simplifying the process of obtaining a view model to an inheritable interface.
This creates the IViewModelConverter interface to be used on provider server data to convert each data set into a child of BaseViewModel with the same method signature.

This also makes the fetching of CFX server data simpler.

This also removes the variable for using the name as the label since we now have status formats, and can hard code those values into the status.